+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/LaunchSettings/LaunchSettingsJsonTemplate.cs
@@ -304,6 +304,34 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.LaunchSettings
                 }
             }
 
+            //Ensure the API profile ends in /swagger or /scalar/v1 esp. if you swap modules.
+            // We dont have a way to know which profiles are for WebApI vs which are for the WebApp for example.
+            foreach (var kvp in launchSettings.Profiles)
+            {
+                var key = kvp.Key;
+                var profile = kvp.Value;
+                if (!key.EndsWith(".Api"))
+                {
+                    continue;
+                }
+                if (!string.IsNullOrWhiteSpace(_defaultLaunchUrlPath) && !profile.LaunchUrl.Contains(_defaultLaunchUrlPath))
+                {
+                    string newPath = _defaultLaunchUrlPath;
+                    if (!newPath.StartsWith("/"))
+                        newPath = "/" + newPath;
+
+                    var uri = new Uri(profile.LaunchUrl);
+
+                    // Build the new URL with the same scheme, host, and port
+                    var builder = new UriBuilder(uri)
+                    {
+                        Path = newPath,
+                        Query = string.Empty // clear any query params if present
+                    };
+                    profile.LaunchUrl = builder.Uri.ToString();
+                }
+            }
+
             if (IsMicrosoftNETSdkWorker())
             {
                 // In case has not been set through an event, sets this be default.